% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{available_database_names}
\alias{available_database_names}
\title{Function provides a vector of available db names}
\usage{
available_database_names()
}
\description{
Function provides a vector of available db names
}
