% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap_db_utilities.R
\name{construct_table_query}
\alias{construct_table_query}
\title{Construct a query, given schema, table, columns, filter condition}
\usage{
construct_table_query(
  schema = "dbo",
  table,
  columns = NULL,
  max_rows = NULL,
  filter_condition = NULL
)
}
\arguments{
\item{schema}{database schema name, string;default="dbo"}

\item{table}{database table name, string}

\item{columns}{character vector of columns; default is NULL, which returns all columns}

\item{max_rows}{maximum number of rows to return; default is NULL (all rows), currently returns TOP max_rows from table}

\item{filter_condition}{string representing 'WHERE' clause of query; default is NULL (no filter)}
}
\value{
query string
}
\description{
Internal function to construct query given user arguments
}
\examples{
construct_table_query(myschema, mytable, col_list, 1000, my_filter)
}
\keyword{internal}
