% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{count}
\alias{count}
\title{Count rows, optionally by group.}
\usage{
count(qry, byvars = NULL, sort = F)
}
\arguments{
\item{qry}{The result set (query/table/view) to return row count for.}

\item{byvars}{optional string vector of column names}

\item{sort}{boolean indicating whether to sort or not (default is F)}
}
\description{
This is a convenience function to count rows of a resultset, optionally by group
}
\examples{
count(qry)
count(qry, byvars=c("var1", "var2"), sort=T)
}
