% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_random_table.R
\name{gen_random_table}
\alias{gen_random_table}
\title{Generate a temp table in the database holding a random set of ids}
\usage{
gen_random_table(
  table,
  idvars,
  schema = "dbo",
  size = 1000,
  filter = NULL,
  seed = NULL,
  engine = default_engine
)
}
\arguments{
\item{table}{string table name to query}

\item{idvars}{vector of string column names that define uniqueness on the table (i.e. the primary key)}

\item{schema}{name of schema (default "dbo")}

\item{size}{number of rows to return (default=1000)}

\item{filter}{string 'WHERE' clause to further restrict the selection}

\item{seed}{integer seed for reproducibility; if no seed is provided, a non-reproducible temp table will be produced}

\item{engine}{a dbConnect connection object; by default will look in namespace for default_engine}
}
\value{
a string name of the temporary table created
}
\description{
This function enables the user to generate a temporary table in the database
that will persist for this connection/session, and that can be utilized in
subsequent queries, in order to consistently restrict to that subset.  This may
be helpful for users to execute an analytic pipeline on a random set of ids.
}
\examples{
rnduids <- gen_random_table(table= "patient", idvars = "osler_id", size = 1000, seed=123)
rnduids <- gen_random_table(table= "patient", idvars = "osler_id", size = 100, seed=87634)
rnduids <- gen_random_table(table= "patient", idvars = "osler_id", filter = "WHERE gender = 'Male'", size = 100, seed=87634)
}
