% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_random_wrk_unit_table.R
\name{gen_random_wrk_unit_table}
\alias{gen_random_wrk_unit_table}
\title{Generate a temp table in the database holding a random set of WRK UNITS}
\usage{
gen_random_wrk_unit_table(
  size = 1000,
  filter = NULL,
  seed = NULL,
  engine = default_engine
)
}
\arguments{
\item{size}{number of unique HOFC_WRK_UNIT_UIDs to return (default 1000)}

\item{filter}{string expressed as a SQL where clause on MHAODS.WRKUNH. The clause must begin with
"WHERE " Examples include:
\itemize{
\item "WHERE MHAODS.WRKUNH.PRC_LVL_CD='H'"
\item "WHERE MHAODS.WRKUNH.PRC_LVL_CD='H' AND MHAODS.WRKUNH.CRNT_HRG_RQSTD_DT>='2017-01-01'"
\item "WHERE PRC_LVL_CD='H' AND CRNT_HRG_RQSTD_DT>='2017-01-01'"
}

The function will return the dbo.tempname of the temporary table (e.g.
"#dplyr_010"); if the results are desired that \code{tempname} can be used in
return_table (i.e. \code{return_table("dbo","#dplyr_010")}). Furthermore, function
\code{return_table()} has an argument \code{random_wrk_units} which can be set
to this \code{tempname} in order to limit the results of that query to this
random subset (See example below).}

\item{seed}{integer seed for reproducibility; if no seed is provided, a non-reproducible temp table will be produced}

\item{engine}{a dbConnect connection object; by default will look in namespace for default_engine}
}
\value{
a string name of the temporary table created
}
\description{
This function enables the user to generate a temporary table in the database
that will persist for this connection/session, and that can be utilized in
subsequent queries, in order to consistently restrict to that subset.  This may
be helpful for users to execute an analytic pipeline on a random set of work
units rather than the full set.
}
\examples{
rnduids <- gen_random_wrk_unit_table(size = 1000, filter="WHERE MHAODS.WRKUNH.PRC_LVL_CD='H'", seed = 123)
rnduids <- gen_random_wrk_unit_table(size = 100, seed=87634)
rnduids <- gen_random_wrk_unit_table(size = 250000, seed=10, engine=myconn)

# Use this handle to the temp table
return_table("dbo",rnduids)

# Use this handle to restrict to another table (for example, pulling all
ARCHWKUT information)
return_table("MHAODS","ARCHWKUT",rnd_wrk_units=rnduids)




}
