% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_random_wrk_unit_table.R
\name{gen_temp_indices}
\alias{gen_temp_indices}
\title{Generate a set of indices}
\usage{
gen_temp_indices(size = 1000, max_index, seed = NULL, engine = default_engine)
}
\arguments{
\item{size}{number of unique HOFC_WRK_UNIT_UIDs to return (default 1000)}
}
\value{
a string name of the temporary table created
}
\description{
This is a helper function, that produces a random set of distinct integers between
1 and maxlength, inclusive.
}
\examples{
indices = gen_temp_indices(size=100000,max_index=14e10, seed=NULL, engine=mycon)
indices = gen_temp_indices(size=1000,max_index=250000, seed=12345, engine=mycon)

}
\keyword{internal}
