% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap_db_utilities.R
\name{get_database_names}
\alias{get_database_names}
\title{Get a list of all databases}
\usage{
get_database_names(pattern = NULL, ...)
}
\arguments{
\item{pattern}{a regular expression to limit searching for dbnames}

\item{...}{passed to get_sql_connection()}
}
\value{
a tibble of database names, id, create_dates for the databases found at the
connection to the server as dictated by \code{...} passed to \code{get_sql_connection()}
}
\description{
This function will connect to "master", and thus does require at least \code{username} or \code{trusted=F}
and will return a list of all db objects, that subsequently might be passed to \code{db_name} of the
get_sql_connection function
}
\examples{
get_database_names(trusted=T)
get_database_names(username="<jhed_id>")
get_database_names(username="<jhed_id>", pattern="(?i)covid")
}
