% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap_db_utilities.R
\name{get_keys}
\alias{get_keys}
\title{Get the primary or foreign keys on a table (if any)}
\usage{
get_keys(
  table,
  schema = "dbo",
  keytype = c("PRIMARY", "FOREIGN"),
  engine = default_engine
)
}
\arguments{
\item{table}{a string name of table to search}

\item{schema}{a string name of schema to search;default="dbo"}

\item{keytype}{one of either 'PRIMARY' (default) or 'FOREIGN'}

\item{engine}{a dbConnect connection object; by default will look in namespace for default_engine}
}
\value{
a character vector of column(s) making up the primary key or foreign key(s) on the table
}
\description{
Function returns a character vector of column name(s) that are the primary or foreign key(s) (see keytype),
given a schema, and table name
}
\examples{
get_keys('encounters', schema="dbo", engine = myconnection)
get_keys('encounters', schema="dbo", keytype='FOREIGN', engine=myconnection)

}
