% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap_db_utilities.R
\name{get_sql_connection}
\alias{get_sql_connection}
\title{Generate a dbConnect object to the SQL database}
\usage{
get_sql_connection(
  dbname,
  username,
  server = "ESMPMDBPR4.WIN.AD.JHU.EDU",
  driver = "FreeTDS",
  tdsver = "8.0",
  verbose = F
)
}
\arguments{
\item{dbname}{string name of database to connect to}

\item{username}{username for the database connection}

\item{server}{defaults to 'ESMPMDBPR4.WIN.AD.JHU.EDU', but another string server could be provided}

\item{driver}{defaults to 'FreeTDS'}

\item{tdsver}{defaults to "8.0"}

\item{verbose}{(logical, default=FALSE; if TRUE will provide some details of the connection)}
}
\description{
This function generates a connection object
}
\examples{
get_sql_connection(dbname = 'CAMP_PMCoE_Projection', username='<jhedid>')
}
