% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap_db_utilities.R
\name{get_view_definition}
\alias{get_view_definition}
\title{Get a view definition}
\usage{
get_view_definition(view, schema = "dbo", engine = default_engine)
}
\arguments{
\item{view}{a string name of the view for which the definition should be returned}

\item{schema}{a string name of schema to search;default is dbo}

\item{engine}{a dbConnect connection object; by default will look in namespace for default_engine}
}
\value{
string of view definition
}
\description{
Function returns the view definition for a view name in a given schema
}
\examples{
get_view_definition("myview", schema="dbo",engine=myconnection)

}
