% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap_db_utilities.R
\name{list_columns}
\alias{list_columns}
\title{List all column names for a given schema, table}
\usage{
list_columns(table, schema = "dbo", engine = default_engine)
}
\arguments{
\item{table}{a string name of table to search}

\item{schema}{a string name of schema to search; default is dbo}

\item{engine}{a dbConnect connection object; by default will look in namespace for default_engine}
}
\value{
character vector of column names
}
\description{
Function returns a character vector of all column names in a table
}
\examples{
list_columns('medications',schema="dbo", engine=myconnection)

}
