% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap_db_utilities.R
\name{list_schemas}
\alias{list_schemas}
\title{List schemas in the database}
\usage{
list_schemas(engine = default_engine)
}
\arguments{
\item{engine}{a dbConnect connection object; by default will look in namespace for default_engine}
}
\value{
character vector of schema names
}
\description{
Function returns a character vector of all schemas in the database
where PRINCIPAL_ID = 1 and where NAME is not 'dbo'; further, if a found
schema has no tables, that schema name is excluded
}
\examples{
list_schemas(myconnection)
}
