% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap_db_utilities.R
\name{list_tables}
\alias{list_tables}
\title{List tables in a given schema}
\usage{
list_tables(schema = "dbo", show_dimensions = FALSE, engine = default_engine)
}
\arguments{
\item{schema}{a string name of schema to search, default="dbo"}

\item{engine}{a dbConnect connection object; by default will look in namespace for default_engine}
}
\value{
character vector of table names
}
\description{
Function returns a character vector of all table names found in a given schema
}
\examples{
list_tables('MIDIB',myconnection)

}
