% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pmap_dbs}
\alias{pmap_dbs}
\title{Function provides a named list of available db names}
\usage{
pmap_dbs(n = NULL)
}
\arguments{
\item{n}{nickname of the projection to return}
}
\description{
Without providing name \code{n}, this function will return
a named list of data base projections available. Provide
argument \code{n} to return just the full name of the specific
projection
}
