% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap_db_utilities.R
\name{query_db}
\alias{query_db}
\title{Query the database with user-defined query string}
\usage{
query_db(query, engine = default_engine)
}
\arguments{
\item{query}{a user-defined query string}

\item{engine}{a dbConnect connection object}
}
\value{
a (lazy) dplyr::tbl / data frame object with results
}
\description{
This function allows a query on the database, using a string. Is equivalent to tbl(con,sql(query)) or dbGetQuery(con,query)
}
\examples{
query_db("SELECT COUNT(*) AS CT FROM MYSCHEMA.MYTABLE", myconnection)
}
