% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap_db_utilities.R
\name{return_table}
\alias{return_table}
\title{Query the database for a specified table, with optional sub-columns/where filter}
\usage{
return_table(
  table,
  schema = "dbo",
  columns = NULL,
  max_rows = NULL,
  filter_condition = NULL,
  rnd_wrk_units = NULL,
  engine = default_engine
)
}
\arguments{
\item{table}{database table name, string}

\item{schema}{database schema name, string; default is "dbo"}

\item{columns}{character vector of columns; default is NULL, which returns all columns}

\item{max_rows}{maximum (integer) number of rows to return; default is NULL
(all rows), currently returns TOP max_rows from table}

\item{filter_condition}{string representing 'WHERE' clause of query; default
is NULL (no filter).}

\item{engine}{a dbConnect connection object}
}
\value{
a (lazy) dplyr::tbl / data frame object with results
}
\description{
This function provides a way to query the database for a single table, which might
be filtered using a where clause, or sub-setting by specified columns/number of rows
}
\examples{
return_table(table="encounters", engine = myconnection)
return_table(table="encounters", columns = c("osler_id"), max_rows=10000, engine = myconnection)
return_table("MHAODS", "STUSHH", columns = c("HOFC_WRK_UNIT_UID", "STUS_CD"), max_rows=10000, filter "year(SRC_SYS_INSRT)=2019", engine = myconnection)
}
