% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap_db_utilities.R
\name{get_sql_connection}
\alias{get_sql_connection}
\title{Generate a dbConnect object to the SQL database}
\usage{
get_sql_connection(
  dbname,
  username,
  server = "ESMPMDBPR4.WIN.AD.JHU.EDU",
  driver = "FreeTDS",
  tdsver = "8.0",
  verbose = F
)
}
\arguments{
\item{dbname}{string name of database to connect to}

\item{username}{username for the database connection}

\item{server}{defaults to 'ESMPMDBPR4.WIN.AD.JHU.EDU', but another string server could be provided}

\item{driver}{defaults to 'FreeTDS'}

\item{tdsver}{defaults to "8.0"}

\item{verbose}{(logical, default=FALSE; if TRUE will provide some details of the connection)}
}
\value{
Database connection object as returned by \code{DBI::dbConnect()}
}
\description{
This function generates a connection object. Note that most of the functions in pmap.utilities
package require the specification of connection engine in the \code{engine} parameter of the function.
However, all functions in the package requiring a connection will look for \code{default_engine} in the
global environment, thus it is convenient (and recommended) to assign the return value of this
to a variable called \code{default_engine}.
}
\examples{
default_engine = get_sql_connection(dbname = 'CAMP_PMCoE_Projection', username='<jhedid>')
}
